Remove extraneous fetch mode
Setting mode=http in the "/tool fetch" command is redundant with using "http://..." in the URL. Also, according to [MikroTik](https://help.mikrotik.com/docs/spaces/ROS/pages/8978514/Fetch) it's also an outdated option. And last but not least, "mode=http" conflicts with the "s" in the https://... of the actual URL being fetched.